requirements.txt
Django==2.2.6
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
nltk==3.4
numpy==1.16.3
pandas==0.24.2
#pkg-resources==0.0.0
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
singledispatch==3.4.0.3
six==1.12.0
pycryptodome
